backend/requirements.txt
# Minimal working version
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3

# AI/ML - Latest compatible versions
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.5

# Development
pytest==7.4.3
httpx==0.25.2